+++ src/main/java/ar/edu/unlam/tallerweb1/modelo/Entrada.java
@@ -9,7 +9,7 @@ public class Entrada implements ItemMenu{
     private Long id;
 
     private String nombre;
-    private Long precio;
+    private Double precio;
     private Double tiempoPreparacion;
     private String descripcion;
 
@@ -37,10 +37,10 @@ public class Entrada implements ItemMenu{
     }
 
     @Override
-    public Long getPrecio() {
+    public Double getPrecio() {
         return precio;
     }
-    public void setPrecio(Long precio) {
+    public void setPrecio(Double precio) {
         this.precio = precio;
     }
