+++ src/main/java/ar/edu/unlam/tallerweb1/modelo/ItemMenu.java
@@ -2,7 +2,7 @@ package ar.edu.unlam.tallerweb1.modelo;
 
 public interface ItemMenu {
     String getNombre();
-    Long getPrecio();
+    Double getPrecio();
     Double getTiempoPreparacion();
     Menu getMenu();
 }
